requirements.txt
numpy>=1.26,<3
PyYAML>=6.0
tqdm>=4.66
rich>=13.7
matplotlib>=3.8
scikit-learn>=1.5
pandas>=2.2
timm>=1.0.8
torchgeo>=0.6.0
rasterio>=1.3.10
shapely>=2.0
geopandas>=1.0.1
earthengine-api>=1.1.0
kornia>=0.7.3
mlflow>=2.17.0
dagshub>=0.3.39
dvc[s3]>=3.56.0
tfrecord>=1.14.4
python-dotenv>=1.0.1
